svm: conformance: use discriminant APIs to drop fd_err_map
The VM error types now expose a `discriminant` method (sbpf #188 for
EbpfError/ElfError, agave #13298 for SyscallError), making the hand-written
variant-to-code tables in `fd_err_map` redundant. Replace them with
`discriminant() + 1` and gather the remaining mapping logic into an `err`
module:

- elf_loader: map ElfError via err::elf_error_code.
- instr/effects: InstructionError has no discriminant method, so keep its
  bincode variant index in err::instruction_error_code.
- syscall: UnpackedResult/unpack_stable_result move into err, mapping
  EbpfError/SyscallError via discriminant.

Note: the discriminant is sequential, so error codes for EbpfError variants
from `InvalidMemoryRegion` onward shift down by one versus the old table,
which skipped index 10 for Firedancer alignment.
